Rename ResultRow → SearchResultRow
