+++ src/components/PlaylistManager/SearchResults/SearchResultRow.js
